Fix all compiler warnings (#405916).
2007-03-29  Matthew Barnes  <mbarnes@redhat.com>

	* camel/camel-exchange-journal.c:
	* camel/camel-exchange-folder.c:
	* camel/camel-stub-marshal.c:
	* camel/camel-exchange-store.c:
	* camel/camel-stub.c:
	* mail/mail-utils.c:
	* mail/mail-utils.h:
	* mail/mail-stub-exchange.c:
	* mail/mail-stub.c:
	* addressbook/e-book-backend-exchange.c:
	* addressbook/e-book-backend-gal.c:
	* calendar/e-cal-backend-exchange.c:
	* calendar/e-cal-backend-exchange-tasks.c:
	* calendar/e-cal-backend-exchange-calendar.c:
	Fix all compiler warnings (#405916).


svn path=/trunk/; revision=1332

+++ addressbook/e-book-backend-gal.c
@@ -2152,7 +2152,7 @@ load_source (EBookBackend *backend,
 			t2 = time (NULL);
 			diff = interval * 24 * 60 *60;
 			printf("Interval:%d, diff:%d\n", interval, diff);
-			printf("t2:%d, t1:%d\n", t2 , t1);
+			printf("t2:%d, t1:%d\n", (int) t2 , (int) t1);
 			/* We have a day specified, then we cache it. */
 			if (!diff || t2 - t1 > diff) {
 				printf ("Cache older than 1 day, refreshing \n");
